enhance: pass edit data from hash insteaed of query string.
